+++ src/components/CodeHighlight/CHANGELOG.md
@@ -7,3 +7,4 @@
 - 2025-10-22: Refactor code. And use JSS for styling.
 - 2025-12-08: Support the highlighting for CSS language, remove the dependency
   on `ThemeProvider`, etc.
+- 2025-12-12: Support the `className` prop to CodeHighlight component.